backend/requirements.txt
Django==4.2
graphene-django==3.1.3
psycopg2-binary==2.9.6
django-cors-headers==4.2.0
Pillow==10.3.0
djangorestframework==3.14.0